backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
passlib[argon2]==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0